requirements.txt
playwright
httpx
beautifulsoup4
python-dotenv
pydantic
iter_tools